requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
pymongo==4.3.3
python-dotenv==1.0.0
Requests==2.31.0
scipy==1.10.1
